requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
paramiko==2.7.2
Pillow==8.2.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
PySimpleGUI==4.39.1
python-dateutil==2.8.1
scipy==1.5.4
six==1.15.0
wincertstore==0.2
